General changes

+++ src/main/java/com/mercadolibre/braavos/invoices/ConversionFactor.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Value
@@ -12,5 +13,5 @@ import java.time.LocalDate;
 public class ConversionFactor {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     LocalDate date;
-    Double value;
+    BigDecimal value;
 }
