+++ src/main/java/com/mercadolibre/braavos/invoices/api/InvoicesSummaryApi.java
@@ -3,11 +3,13 @@ package com.mercadolibre.braavos.invoices.api;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @AllArgsConstructor
 public class InvoicesSummaryApi {
     String userId;
-    Double totalCharges;
-    Double totalPayments;
-    Double total;
+    BigDecimal totalCharges;
+    BigDecimal totalPayments;
+    BigDecimal total;
 }
