+++ src/main/java/com/mercadolibre/braavos/invoices/payments/PaymentInputApi.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Value;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Value
 @Builder(toBuilder = true)
@@ -16,5 +17,5 @@ public class PaymentInputApi {
     @NotNull
     String currency;
     @NotNull
-    Double amount;
+    BigDecimal amount;
 }
